requirements.txt
epicscorelibs
epics-pypdb>=0.1.5
jinja2
graphviz
